move func back in to file
