+++ src/projectGeneration/templates.ts
@@ -1,8 +1,17 @@
 import * as vscode from "vscode";
 import { ScaffoldV1Template, ScaffoldV1TemplateSpec } from "../clients/scaffoldingService";
 import { QuickPickItemWithValue } from "../quickpicks/types";
-import { filterSensitiveKeys } from "../scaffoldUtil";
 import { getSidecar } from "../sidecar";
+
+export function filterSensitiveKeys<T>(obj: Record<string, T>): Record<string, T> {
+  return Object.fromEntries(
+    Object.entries(obj)
+      .filter(
+        ([key]) => !key.toLowerCase().includes("key") && !key.toLowerCase().includes("secret"),
+      )
+      .map(([key, value]) => [key, value]),
+  );
+}
 
 export function sanitizeTemplateOptions(template: ScaffoldV1Template): ScaffoldV1Template {
   const spec = template.spec as ScaffoldV1TemplateSpec;
