--- src/scaffoldUtil.test.ts
@@ -1,24 +1,0 @@
-import * as assert from "assert";
-import { filterSensitiveKeys } from "./scaffoldUtil";
-
-describe("filterSensitiveKeys", () => {
-  it("filters out keys containing 'key' or 'secret'", () => {
-    const input = {
-      api_key: "sensitive",
-      secret_token: "sensitive",
-      bootstrap_server: "localhost:9092",
-      topic_name: "test-topic",
-    };
-
-    const result = filterSensitiveKeys(input);
-
-    assert.deepStrictEqual(
-      result,
-      {
-        bootstrap_server: "localhost:9092",
-        topic_name: "test-topic",
-      },
-      "Should filter out sensitive keys while preserving others",
-    );
-  });
-});
